Reorganize where layouts live
