+++ src/components/Layout/index.js
@@ -1,0 +1,6 @@
+import Layout from './Layout'
+
+export { default as ContentPageLayout } from './subLayouts/ContentPageLayout'
+export { default as HomePageLayout } from './subLayouts/HomePageLayout'
+
+export default Layout
