+++ src/components/pages/Collect/Collect.js
@@ -3,6 +3,7 @@ import { toast } from 'react-toastify'
 import { usePagination, useSortBy, useTable } from 'react-table'
 import React, { useEffect, useMemo, useState } from 'react'
 
+import { ContentPageLayout } from '../../Layout'
 import { databaseSwitchboardPropTypes } from '../../../App/mermaidData/databaseSwitchboard/DatabaseSwitchboard'
 import { H3 } from '../../generic/text'
 import { reactTableNaturalSort } from '../../generic/Table/reactTableNaturalSort'
@@ -12,7 +13,6 @@ import AddSampleUnitButton from './AddSampleUnitButton'
 import language from '../../../language'
 import PageSelector from '../../generic/Table/PageSelector'
 import PageSizeSelector from '../../generic/Table/PageSizeSelector'
-import ContentPageLayout from '../../ContentPageLayout'
 import useCurrentProjectPath from '../../../library/useCurrentProjectPath'
 
 const TopBar = () => (
