+++ src/components/pages/ManagementRegimes/ManagementRegimes.js
@@ -1,6 +1,6 @@
 import React from 'react'
 // import PropTypes from 'prop-types'
-import ContentPageLayout from '../../ContentPageLayout'
+import { ContentPageLayout } from '../../Layout'
 
 const ManagementRegimes = () => {
   return (
