+++ src/components/pages/Projects/Projects.js
@@ -2,11 +2,11 @@ import { toast } from 'react-toastify'
 import React, { useEffect, useState } from 'react'
 
 import { databaseSwitchboardPropTypes } from '../../../App/mermaidData/databaseSwitchboard'
+import { HomePageLayout } from '../../Layout'
 import language from '../../../language'
 import LoadingIndicator from '../../LoadingIndicator/LoadingIndicator'
 import ProjectCard from '../../ProjectCard'
 import ProjectToolBarSection from '../../ProjectToolBarSection'
-import HomePageLayout from '../../HomePageLayout'
 
 /**
  * All Projects page (lists projects)
