+++ src/components/pages/Users/Users.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import ContentPageLayout from '../../ContentPageLayout'
+import { ContentPageLayout } from '../../Layout'
 // import PropTypes from 'prop-types'
 
 const Users = () => {
